Factor out ds_factory_info structure.
